ultimos cambios

+++ platforms/android/assets/www/js/controllers.js
@@ -6,6 +6,7 @@ angular.module('starter.controllers', [])
 	$scope.Guardar=function(){
 		var dato=JSON.stringify($scope.usuarioLog);
 		$state.go("tab.piano",{nombre:dato});
+		$scope.usuarioLog.nombreLog="";
 	}
 })
 
@@ -15,14 +16,15 @@ angular.module('starter.controllers', [])
 		  $cordovaFile.createFile(cordova.file.externalRootDirectory, "piano.txt",true) // cordova.file.dataDirectory //cordova.file.externalRootDirectory
 	      .then(function (success) {
 	        // success
+	        console.log("archivo creado");
 	      }, function (error) {
 	        // error
+	        console.log(error);
 	      });
 		}, false);
 
 	$scope.usuario={};
 	var nombre=JSON.parse($stateParams.nombre);
-  	$scope.usuario.puntaje=0;
   	$scope.usuario.nombre=nombre.nombreLog;
 
 	$scope.botones = [1,2,3,4,5,6,7,8,9];
@@ -30,19 +32,27 @@ angular.module('starter.controllers', [])
 	$scope.filas = $scope.botones.length / $scope.columnas; // agrupo los botones de a 3
 	$scope.grabados=[];
 
-	$cordovaFile.checkFile(cordova.file.externalRootDirectory, "piano.txt") // cordova.file.dataDirectory //cordova.file.externalRootDirectory
-        .then(function (success) {
-          // succes
-          $cordovaFile.readAsText(cordova.file.externalRootDirectory, "piano.txt")
-                  .then(function (success) {
-                    var dato=JSON.parse(success);
-                    $scope.grabados=dato;
-                  }, function (error) {
-                    // error
-                  });
-        }, function (error) {
-          // error
-        });
+	try{
+		$cordovaFile.checkFile(cordova.file.externalRootDirectory, "piano.txt") // cordova.file.dataDirectory //cordova.file.externalRootDirectory
+	        .then(function (success) {
+	          // succes
+	          $cordovaFile.readAsText(cordova.file.externalRootDirectory, "piano.txt")
+	                  .then(function (success) {
+	                    var dato=JSON.parse(success);
+	                    $scope.grabados=dato;
+	                  }, function (error) {
+	                    // error
+	                    console.log(error);
+	                  });
+	        }, function (error) {
+	          // error
+	          console.log(error);
+	        });
+	}
+	catch(e)
+	{
+		console.log("El plugin File funciona en dispositivos unicamente");
+	}
 
 	$scope.getFilas = function(){
 	   return new Array($scope.filas);
@@ -57,15 +67,7 @@ angular.module('starter.controllers', [])
 		    	{
 		    		$scope.apretado.push('bulbasaur');
 		    	}
-			   	try
-			  	{
-					$cordovaVibration.vibrate(300);
-		    		$cordovaNativeAudio.play('bulbasaur');
-		    	}
-		    	catch(e)
-		    	{
-		    		console.log("La vibracion y el sonido, solo funcionan en celulares");
-		    	}
+		    	Reproducir('bulbasaur',300);
 		    	break;
 
 		    case 'btn_fila_0_col_1':
@@ -73,15 +75,7 @@ angular.module('starter.controllers', [])
 		    	{
 		    		$scope.apretado.push('charmander');
 		    	}
-			   	try
-			  	{
-					$cordovaVibration.vibrate(300);
-		    		$cordovaNativeAudio.play('charmander');
-		    	}
-		    	catch(e)
-		    	{
-		    		console.log("La vibracion y el sonido, solo funcionan en celulares");
-		    	}
+		    	Reproducir('charmander',300);
 		    	break;
 
 	   		case 'btn_fila_0_col_2':
@@ -89,15 +83,7 @@ angular.module('starter.controllers', [])
 		    	{
 		   			$scope.apretado.push('clefairy');
 		   		}
-			   	try
-			  	{
-					$cordovaVibration.vibrate(300);
-		    		$cordovaNativeAudio.play('clefairy');
-		    	}
-		    	catch(e)
-		    	{
-		    		console.log("La vibracion y el sonido, solo funcionan en celulares");
-		    	}
+		   		Reproducir('clefairy',300);
 		    	break;
 
 	   		case 'btn_fila_1_col_0':
@@ -105,15 +91,7 @@ angular.module('starter.controllers', [])
 		    	{
 		   			$scope.apretado.push('eevee');
 		   		}
-			   	try
-			  	{
-					$cordovaVibration.vibrate(300);
-		    		$cordovaNativeAudio.play('eevee');
-		    	}
-		    	catch(e)
-		    	{
-		    		console.log("La vibracion y el sonido, solo funcionan en celulares");
-		    	}
+		   		Reproducir('eevee',300);
 		    	break;
 
 	   		case 'btn_fila_1_col_1':
@@ -121,15 +99,7 @@ angular.module('starter.controllers', [])
 		    	{
 		   			$scope.apretado.push('meowth');
 		   		}
-			   	try
-			  	{
-					$cordovaVibration.vibrate(300);
-		    		$cordovaNativeAudio.play('meowth');
-		    	}
-		    	catch(e)
-		    	{
-		    		console.log("La vibracion y el sonido, solo funcionan en celulares");
-		    	}
+		   		Reproducir('meowth',300);
 		    	break;
 
 	   		case 'btn_fila_1_col_2':
@@ -137,15 +107,7 @@ angular.module('starter.controllers', [])
 		    	{
 		   			$scope.apretado.push('pikachu');
 		   		}
-			   	try
-			  	{
-					$cordovaVibration.vibrate(300);
-		    		$cordovaNativeAudio.play('pikachu');
-		    	}
-		    	catch(e)
-		    	{
-		    		console.log("La vibracion y el sonido, solo funcionan en celulares");
-		    	}
+		   		Reproducir('pikachu',300);
 		    	break;
 
 	   		case 'btn_fila_2_col_0':
@@ -153,15 +115,7 @@ angular.module('starter.controllers', [])
 		    	{
 		   			$scope.apretado.push('psyduck');
 		   		}
-			   	try
-			  	{
-					$cordovaVibration.vibrate(300);
-		    		$cordovaNativeAudio.play('psyduck');
-		    	}
-		    	catch(e)
-		    	{
-		    		console.log("La vibracion y el sonido, solo funcionan en celulares");	    	
-		    	}
+		   		Reproducir('psyduck',300);
 		    	break;
 
 	   		case 'btn_fila_2_col_1':
@@ -169,15 +123,7 @@ angular.module('starter.controllers', [])
 		    	{
 		   			$scope.apretado.push('butterfree');
 		   		}
-			   	try
-			  	{
-					$cordovaVibration.vibrate(300);
-		    		$cordovaNativeAudio.play('butterfree');
-		    	}
-		    	catch(e)
-		    	{
-		    		console.log("La vibracion y el sonido, solo funcionan en celulares");
-		    	}
+				Reproducir('butterfree',300);
 		    	break;
 
 	   		case 'btn_fila_2_col_2':
@@ -185,17 +131,21 @@ angular.module('starter.controllers', [])
 		    	{
 		   			$scope.apretado.push('squirtle');
 		   		}
-			   	try
-			  	{
-					$cordovaVibration.vibrate(300);
-		    		$cordovaNativeAudio.play('squirtle');
-		    	}
-		    	catch(e)
-		    	{
-		    		console.log("La vibracion y el sonido, solo funcionan en celulares");
-		    	}
+		   		Reproducir('squirtle',300);
 		    	break;
     	} 
+	}
+
+	function Reproducir(sonido,tiempo){
+		try
+		{
+			$cordovaVibration.vibrate(tiempo);
+			$cordovaNativeAudio.play(sonido);
+		}
+		catch(e)
+		{
+			console.log("La vibracion y el sonido, solo funcionan en celulares");
+		}
 	}
 
 	$scope.Grabar=function(){
@@ -207,27 +157,37 @@ angular.module('starter.controllers', [])
 		archivo.nombre=$scope.usuario.nombre;
 		archivo.sonidos=$scope.apretado;
 
-		var confirmPopup = $ionicPopup.confirm({
-     		title: 'Guardar',
-     		template: '¿Desea guardar en archivo?',
-     		cssClass:'salida',
-     		okType: 'button-energized',
-   		});
+		if($scope.apretado!=null && $scope.apretado.length!=0)
+		{
+			var confirmPopup = $ionicPopup.confirm({
+	     		title: 'Guardar',
+	     		template: '¿Desea guardar en archivo?',
+	     		cssClass:'salida',
+	     		okType: 'button-energized',
+	   		});
+	   		
+	      $scope.grabados.push(archivo);
+	      var dato=JSON.stringify($scope.grabados);
 
-      $scope.grabados.push(archivo);
-      var dato=JSON.stringify($scope.grabados);
-
-   		confirmPopup.then(function(res) {
-     		if(res) {
-       			$cordovaFile.writeFile(cordova.file.externalRootDirectory, "piano.txt", dato, true)
-				      .then(function (success) {
-				        // success
-				        console.log("archivo guardado");
-				      }, function (error) {
-				        // error
-				      });
-     		}
-   		});
+	   		confirmPopup.then(function(res) {
+	     		if(res) {
+	     			try{     				
+		       			$cordovaFile.writeFile(cordova.file.externalRootDirectory, "piano.txt", dato, true)
+						      .then(function (success) {
+						        // success
+						        console.log("archivo guardado");
+						      }, function (error) {
+						        // error
+						        console.log(error);
+						      });
+					}
+					catch(e)
+					{
+						console.log("El plugin File funciona en dispositivos unicamente");
+					}
+	     		}
+	   		});
+	   	}
 
 	    $scope.copia=$scope.apretado;
 	    $scope.apretado=null;
@@ -260,9 +220,10 @@ angular.module('starter.controllers', [])
 				      .then(function (success) {
 				        // success
 				        var dato=JSON.parse(success);
-				        $scope.archivo=dato;
+				        $scope.archivo=dato;		                
 				      }, function (error) {
 				        // error
+				        console.log(error);
 				      });
 		}, false);
 })
